refactor: qdraw をモジュール分割して責務ごとに整理
