+++ script/lib/svg_builder.rb
@@ -1,0 +1,15 @@
+module QDraw
+  class SvgBuilder
+    def initialize
+      @parts = []
+    end
+
+    def add(text)
+      @parts << text
+    end
+
+    def to_s
+      @parts.join("\n")
+    end
+  end
+end
